Refactoring main scripts and nce module
Major changes are:
    - The parser setup is moved to utils.
    - index_module is now built outside NCELoss, for the portability
    - IndexLinear module is now moved to separate file
