+++ main.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
 import sys
-import argparse
 import time
 from datetime import datetime
 import math
@@ -12,57 +11,9 @@ import torch.optim as optim
 import data
 from model import RNNModel
 from nce import NCELoss
-from utils import process_data, build_unigram_noise
-
-def setup_parser():
-    parser = argparse.ArgumentParser(
-        description='PyTorch PennTreeBank RNN/LSTM Language Model')
-    parser.add_argument('--data', type=str, default='./data/penn',
-                        help='location of the data corpus')
-    parser.add_argument('--dict', type=str, default=None,
-                        help='location of the vocabulary file, without which will use vocab of training corpus')
-    parser.add_argument('--emsize', type=int, default=200,
-                        help='size of word embeddings')
-    parser.add_argument('--nhid', type=int, default=200,
-                        help='number of hidden units per layer')
-    parser.add_argument('--nlayers', type=int, default=2,
-                        help='number of layers')
-    parser.add_argument('--lr', type=float, default=1.0,
-                        help='initial learning rate')
-    parser.add_argument('--weight-decay', type=float, default=1e-5,
-                        help='initial weight decay')
-    parser.add_argument('--lr-decay', type=float, default=2,
-                        help='learning rate decay when no progress is observed on validation set')
-    parser.add_argument('--clip', type=float, default=0.25,
-                        help='gradient clipping')
-    parser.add_argument('--epochs', type=int, default=40,
-                        help='upper epoch limit')
-    parser.add_argument('--batch-size', type=int, default=20, metavar='N',
-                        help='batch size')
-    parser.add_argument('--dropout', type=float, default=0.2,
-                        help='dropout applied to layers (0 = no dropout)')
-    parser.add_argument('--seed', type=int, default=1111,
-                        help='random seed')
-    parser.add_argument('--cuda', action='store_true',
-                        help='use CUDA')
-    parser.add_argument('--log-interval', type=int, default=200, metavar='N',
-                        help='report interval')
-    parser.add_argument('--save', type=str, default='model.pt',
-                        help='path to save the final model')
-    parser.add_argument('--nce', action='store_true',
-                        help='use NCE as loss function')
-    parser.add_argument('--noise-ratio', type=int, default=10,
-                        help='set the noise ratio of NCE sampling')
-    parser.add_argument('--norm-term', type=int, default=9,
-                        help='set the log normalization term of NCE sampling')
-    parser.add_argument('--train', action='store_true',
-                        help='set train mode, otherwise only evaluation is performed')
-    parser.add_argument('--tb-name', type=str, default=None,
-                        help='the name which would be used in tensorboard record')
-    parser.add_argument('--prof', action='store_true',
-                        help='Enable profiling mode, will execute only one batch data')
-    return parser
-
+from utils import process_data, build_unigram_noise, setup_parser
+from index_gru import IndexGRU
+from index_linear import IndexLinear
 
 parser = setup_parser()
 args = parser.parse_args()
@@ -102,17 +53,18 @@ eval_batch_size = args.batch_size
 ################################################################## Build the criterion and model
 #################################################################
 
-ntokens = len(corpus.train.dataset.dictionary)
-print('Vocabulary size is {}'.format(ntokens))
+ntoken = len(corpus.train.dataset.dictionary)
+print('Vocabulary size is {}'.format(ntoken))
 
 # noise for soise sampling in NCE
 noise = build_unigram_noise(
     torch.FloatTensor(corpus.train.dataset.dictionary.idx2count)
 )
 
+index_module = IndexLinear(args.nhid, ntoken)
+#index_module = IndexGRU(ntoken, nhidden, nhidden, 0.5)
 criterion = NCELoss(
-    ntokens=ntokens,
-    nhidden=args.nhid,
+    index_module=index_module,
     noise=noise,
     noise_ratio=args.noise_ratio,
     norm_term=args.norm_term,
@@ -120,7 +72,7 @@ criterion = NCELoss(
 criterion.nce_mode(args.nce)
 
 model = RNNModel(
-    ntokens, args.emsize, args.nhid, args.nlayers,
+    ntoken, args.emsize, args.nhid, args.nlayers,
     criterion=criterion, dropout=args.dropout,
 )
 if args.cuda:
